исправлен мелкий недочет

+++ Лаба 3 Вариант Б5/asset.cpp
@@ -12,7 +12,7 @@
 
 
 
-Asset::Asset(const std::string & _FullNameProperty, int _cost)
+Asset::Asset(const std::string & _FullNameProperty, double _cost)
 	:m_FullNameProperty(_FullNameProperty),
 	m_cost(_cost)
 {
